Breaking change to type system, deprecated more
The following types are now changed:
- * is now addr
- & is now ref
- [] is now array
- <> is now template
- && is deprecated

I did this because I believe symbol-heavy languages end up harder to
read and understand. Now, we can do things that make a lot of sense:
(var my-int-ptr (addr int) (addr an-int))
It also means less utilization of the shift key, which is nice
ergonomics-wise. The tradeoff of course is more typing, but I consider
pressing a single key quite a bit easier than shift and a key.

The following generators are now deprecated:
++
--
%
block

The first three are in the same spirit as the type name
change. Symbols are harder to search and type than words.

These are now renamed:
bit->> becomes bit-shift->>
bit-<< becomes bit-shift-<<

This is an obnoxious change, but I figure because of low utilization
by people other than myself, now is as good a time as any.