Added ability to defer macro resolution
Gamelib's XML.cake depends on the XML format to be defined before the
write-xml-in-format macro can function. However, this wasn't possible
in some cases because the write-xml-in-format was being evaluated
before define-xml-format.

To fix this, I added a non-ideal thing. If your macro needs to try to
wait longer before resolving, you can call
DeferCurrentReferenceResolution in the macro.

The evaluator then saves the reference off until all other references
are resolved, then re-evaluates the deferred reference.