Hacked in C mode, which will break existing code
This is in support of moving more towards C by default rather than C++
for final program output.

* Pass an "exposure" argument which informs the system whether it
needs to consider the C++ requirement global or confined to the
module. Currently there is no infection, so if a C module imports a
one with C++ in the header, things will break.
* Add stdbool to make C compilation work in various tests
* Forward-declare functions in HotReloadingCodeModifier.cake rather
than importing their modules
* Use NULL rather than fooling around with nullptr swapping
* Properly require C++ on definitions with types that use C++ features
* Properly determine comptime vs. build time code and only pay
attention to build time requirements when determining which compiler
to use
* Hack in the use of gcc over g++ when the module is determined to be
C compatible. I'm going to revisit this very soon.